LayoutTests/platform/chromium-cg-mac-snowleopard/http/tests/security/mixedContent/redirect-http-to-https-video-in-main-frame-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
CONSOLE MESSAGE: line 1: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-redirect-http-to-https-video.html displayed insecure content from http://127.0.0.1:8080/security/resources/redir.php?url=https://127.0.0.1:8443/resources/test.ogv.

didDisplayInsecureContent
CONSOLE MESSAGE: line 1: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-redirect-http-to-https-video.html displayed insecure content from http://127.0.0.1:8080/security/resources/redir.php?url=https://127.0.0.1:8443/resources/test.ogv.

didDisplayInsecureContent
This test opens a window that loads an insecure video(via a tricky redirect). We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
